origin/main carried an independent hardening pass (commit 41911c4, ~34
files) done against the 2026-05-10 audit while this branch fixed the
2026-06-06 audit on the older 1ae453b base. Both are kept -- the union,
not one side winning. 9 files conflicted; resolved as:

- WorldCongress.hpp: identical fix; kept ours (adds the static_assert).
- WorldEvents.cpp / GameDBus.cpp: same fix both sides; kept the richer
  variant (explicit cast; LOG_ERROR + deactivate the dead bus).
- Combat.cpp: both fixed the ranged-combat UAF differently; kept our
  PendingKill list (mirrors the melee path), dropped remote's parallel
  defKillPlayer so no orphaned var remained.
- AIController.cpp: complementary -- our disarm-on-failure (the audit
  finding) PLUS remote's defensive `launcher = nullptr` after the blast.
- TurnProcessor.cpp: both fixed the atWar OOB; kept our diagnostics
  (static_assert + LOG_WARN) and adopted remote's extra outer-index
  guard (i < atWarCount), using the single canonical atWarCount.
- DebugServer.cpp: kept remote's superior JSON escaping (control chars +
  \u00XX) but with our public string_view signature (Application.cpp
  calls it) instead of remote's file-local const char* version.
- LuaEngine.cpp: union -- our text-only chunk loader/wrappers PLUS
  remote's instruction-limit watchdog and post-init sandbox verification.
- Serializer.cpp: kept our canReadRecords-based caps + dimension checks
  for all 13 overlapping sites; removed the 4 duplicate cap-constant
  definitions (kept the 2026-05-10 block's) and the redundant second
  map-dimension check. Remote's skip()/hasRemaining bounds hardening
  (non-conflicting) is preserved.

Verified: release build clean; ctest 12/12; determinism byte-identical;
golden re-blessed (merge integrates remote's gameplay changes, so sim
results shift once -- expected).

+++ src/simulation/unit/Movement.cpp
@@ -227,8 +227,8 @@ bool moveUnitAlongPath(aoc::game::GameState& gameState, aoc::game::Unit& unit,
                                 static_cast<int32_t>(city->walls().currentHP),
                                 m.combatBonusVsCities);
                             if (extra > 0) {
-                                city->walls().currentHP = static_cast<int16_t>(
-                                    city->walls().currentHP - extra);
+                                city->walls().currentHP =
+                                    city->walls().currentHP - extra;
                                 siegeDmg += extra;
                             }
                         }
